+++ app/src/main/res/layout/activity_main.xml
@@ -24,5 +24,10 @@
         <include layout="@layout/search_main" />
     </LinearLayout>
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+        <include layout="@layout/route_card_compact" />
+    </LinearLayout>
 </RelativeLayout>
