+++ app/src/main/res/layout/route_card_compact.xml
@@ -14,6 +14,7 @@
         android:orientation="vertical"
         app:behavior_hideable="false"
         app:behavior_peekHeight="60dp"
+        android:background="@color/white"
         app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
 
         <RelativeLayout
